Fix mkTemp, et al. to handle the case where an X appears in the dir part of the template.
This happens quite often on Mac OS X in my experience.  $TMPDIR is set to a private location with the name apparently generated by some hashing algorithm.  I wouldn't be surprised if it is derived from the username somehow, as my accounts consistently wind up with 'X's in their $TMPDIR, on many systems.